pyproject.toml
[project]
name = "personalbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Akromjon"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular-sidecar (>=2025.2.1,<2026.0.0)",
    "django (>=5.1.6,<6.0.0)",
    "django (>=5.1.6,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-modeltranslation (>=0.19.12,<0.20.0)",
    "django-ckeditor-5 (>=0.2.17,<0.3.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-rosetta (>=0.10.1,<0.11.0)",
    "django-unfold (>=0.49.1,<0.50.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "django-silk (>=5.3.2,<6.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "aiogram (>=3.18.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
